master_pi/requirements.txt
pyserial>=3.5
RPi.GPIO>=0.7.1
paho-mqtt>=1.6.1
